feat(admin): split File Upload Test into dedicated page with tabbed views
- Add ADMIN_PORTAL_TEST_PATH (env-driven, dev fallback /system/admin/test)
  and mount AdminTestPage under it
- Lift listMyFileCdnUrls into useMyFileCdnUrls hook so view switches share
  one fetch instead of remounting per tab
- Split UI into FileUploadView (upload only) and MyFilesView (full URL list
  + Rendering top 3) toggled via ?view=upload|files chip tabs
- Limit Rendering preview to 3 most recent files; full list stays visible
  in My File URLs above
- Restructure AdminProfilePage: add 테스트 기능 section with File Upload
  Test entry card and API 헬스체크 (출시 예정) placeholder, remove the
  obsolete inline AdminFileUploadTestSection
