+++ src/App.tsx
@@ -25,10 +25,12 @@ import { ArticleEditPage } from '@/pages/Articles/ArticleEditPage';
 import { AdminLoginPage } from '@/pages/AdminPortal/AdminLoginPage';
 import { AdminSignupPage } from '@/pages/AdminPortal/AdminSignupPage';
 import { AdminProfilePage } from '@/pages/AdminPortal/AdminProfilePage';
+import { AdminTestPage } from '@/pages/AdminPortal/AdminTestPage';
 import {
     ADMIN_PORTAL_LOGIN_PATH,
     ADMIN_PORTAL_PROFILE_PATH,
     ADMIN_PORTAL_SIGNUP_PATH,
+    ADMIN_PORTAL_TEST_PATH,
 } from '@/constants/adminPortal';
 
 function ScrollToTopOnPathChange() {
@@ -75,6 +77,7 @@ export const App = () => {
                         <Route path="/profile" element={<UserProfileRoute />} />
                         <Route element={<AdminRoute />}>
                             <Route path={ADMIN_PORTAL_PROFILE_PATH} element={<AdminProfilePage />} />
+                            <Route path={ADMIN_PORTAL_TEST_PATH} element={<AdminTestPage />} />
                             <Route path="/articles/create" element={<ArticleCreatePage />} />
                             <Route path="/articles/:id/edit" element={<ArticleEditPage />} />
                         </Route>
